requirements.txt
astroid==2.2.5
boto3==1.9.227
botocore==1.12.227
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
docutils==0.15.2
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
Werkzeug==0.15.5
wrapt==1.11.2
